Simplify stm_task_queue_register steps
Since coroutine and stm_TaskSplit functions are added,
the task register should be redesigned and clear.

In the modification, we used three steps to figure out the task
queue states.
(1) Allocated with thread_id
(2) Allocated without thread_id
(3) Not allocated

+++ include/task_queue.h
@@ -89,9 +89,12 @@ ws_task_circular_array_new(unsigned long size)
 static inline void
 ws_task_circular_array_delete(ws_task_circular_array* ws_array)
 {
+    /*
     for (long i=0; i < ws_array->_size; i++) {
-        free(ws_array->_array[(int)i]);
+        free(ws_array->_array[i]);
     }
+    */
+    
     free(ws_array->_array);
     free(ws_array);
 }
